LPD-95511 Move the import deadlock test to the executor test class
