+++ modules/apps/layout/layout-set-prototype-test/src/testIntegration/java/com/liferay/layout/set/prototype/exportimport/test/LayoutSetPrototypeExportImportTest.java
@@ -8,7 +8,6 @@ package com.liferay.layout.set.prototype.exportimport.test;
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.exportimport.kernel.configuration.ExportImportConfigurationSettingsMapFactoryUtil;
 import com.liferay.exportimport.kernel.configuration.constants.ExportImportConfigurationConstants;
-import com.liferay.exportimport.kernel.lar.ExportImportThreadLocal;
 import com.liferay.exportimport.kernel.model.ExportImportConfiguration;
 import com.liferay.exportimport.kernel.service.ExportImportConfigurationLocalServiceUtil;
 import com.liferay.exportimport.kernel.service.ExportImportLocalServiceUtil;
@@ -16,7 +15,6 @@ import com.liferay.exportimport.test.util.lar.BasePortletExportImportTestCase;
 import com.liferay.layout.set.prototype.constants.LayoutSetPrototypePortletKeys;
 import com.liferay.layout.test.util.LayoutTestUtil;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.LayoutPrototype;
 import com.liferay.portal.kernel.model.LayoutSetPrototype;
@@ -28,15 +26,10 @@ import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.test.util.UserTestUtil;
-import com.liferay.portal.kernel.transaction.Propagation;
-import com.liferay.portal.kernel.transaction.TransactionConfig;
-import com.liferay.portal.kernel.transaction.TransactionInvokerUtil;
-import com.liferay.portal.kernel.util.Time;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 
 import java.io.File;
 
-import java.util.Date;
 import java.util.Map;
 
 import org.junit.Assert;
@@ -167,76 +160,6 @@ public class LayoutSetPrototypeExportImportTest
 		exportImportLayoutSetPrototype(true);
 	}
 
-	@Test
-	@TestInfo("LPD-95511")
-	public void testImportLayoutSetPrototypeWithLayout() throws Exception {
-		LayoutSetPrototype layoutSetPrototype =
-			LayoutTestUtil.addLayoutSetPrototype(RandomTestUtil.randomString());
-
-		Group group = layoutSetPrototype.getGroup();
-
-		long layoutSetPrototypeId =
-			layoutSetPrototype.getLayoutSetPrototypeId();
-
-		boolean layoutImportInProcess =
-			ExportImportThreadLocal.isLayoutImportInProcess();
-
-		ExportImportThreadLocal.setLayoutImportInProcess(true);
-
-		try {
-			TransactionInvokerUtil.invoke(
-				_requiredTransactionConfig,
-				() -> {
-
-					// Lock the layout set prototype row in the outer
-					// transaction, as importing its staged model does
-
-					LayoutSetPrototype lockedLayoutSetPrototype =
-						LayoutSetPrototypeLocalServiceUtil.
-							getLayoutSetPrototype(layoutSetPrototypeId);
-
-					Date modifiedDate =
-						lockedLayoutSetPrototype.getModifiedDate();
-
-					long time = modifiedDate.getTime();
-
-					lockedLayoutSetPrototype.setModifiedDate(
-						new Date(time - Time.DAY));
-
-					LayoutSetPrototypeLocalServiceUtil.updateLayoutSetPrototype(
-						lockedLayoutSetPrototype);
-
-					// Import a content page in a nested transaction, as the
-					// batch engine does for each page
-
-					try {
-						TransactionInvokerUtil.invoke(
-							_requiresNewTransactionConfig,
-							() -> LayoutTestUtil.addTypeContentLayout(
-								group, true, false));
-					}
-					catch (Throwable throwable) {
-						throw new SystemException(throwable);
-					}
-
-					return null;
-				});
-		}
-		catch (Throwable throwable) {
-			throw new AssertionError(
-				"Importing a content page into a layout set prototype " +
-					"deadlocked",
-				throwable);
-		}
-		finally {
-			ExportImportThreadLocal.setLayoutImportInProcess(
-				layoutImportInProcess);
-
-			LayoutSetPrototypeLocalServiceUtil.deleteLayoutSetPrototype(
-				layoutSetPrototypeId);
-		}
-	}
-
 	protected void exportImportLayoutSetPrototype(boolean layoutPrototype)
 		throws Exception {
 
@@ -305,12 +228,5 @@ public class LayoutSetPrototypeExportImportTest
 	protected void testExportImportDisplayStyle(long groupId, String scopeType)
 		throws Exception {
 	}
-
-	private static final TransactionConfig _requiredTransactionConfig =
-		TransactionConfig.Factory.create(
-			Propagation.REQUIRED, new Class<?>[] {Exception.class});
-	private static final TransactionConfig _requiresNewTransactionConfig =
-		TransactionConfig.Factory.create(
-			Propagation.REQUIRES_NEW, new Class<?>[] {Exception.class});
 
 }
